More groups, less files
