+++ InnateMC/Data/Java/DownloadedJavaInstallation.swift
