+++ InnateMC/Data/Java/JreDownloader.swift
