+++ InnateMC/Data/Java/LinkedJavaInstallation.swift
